Reorganize conversation list into shared, chats and archived packages
