+++ AndroidManifest.xml
@@ -71,7 +71,7 @@
 
         <!-- Displays a list of conversations -->
         <activity
-            android:name=".ui.conversationlist.ConversationListActivity"
+            android:name=".ui.conversationlist.chats.ConversationListActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:screenOrientation="user"
             android:exported="true"
@@ -117,10 +117,10 @@
             android:screenOrientation="user"
             android:label="@string/archived_activity_title"
             android:theme="@style/Theme.Compose"
-            android:parentActivityName="com.android.messaging.ui.conversationlist.ConversationListActivity">
+            android:parentActivityName="com.android.messaging.ui.conversationlist.chats.ConversationListActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.android.messaging.ui.conversationlist.ConversationListActivity" />
+                android:value="com.android.messaging.ui.conversationlist.chats.ConversationListActivity" />
         </activity>
 
         <!-- Displays the contents of a single conversation -->
@@ -132,10 +132,10 @@
             android:resizeableActivity="true"
             android:windowSoftInputMode="stateHidden|adjustResize"
             android:theme="@style/Theme.Compose"
-            android:parentActivityName="com.android.messaging.ui.conversationlist.ConversationListActivity">
+            android:parentActivityName="com.android.messaging.ui.conversationlist.chats.ConversationListActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.android.messaging.ui.conversationlist.ConversationListActivity" />
+                android:value="com.android.messaging.ui.conversationlist.chats.ConversationListActivity" />
         </activity>
 
         <!-- Blocked Participants -->
@@ -144,13 +144,13 @@
             android:allowEmbedded="true"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:label="@string/blocked_contacts_title"
-            android:parentActivityName="com.android.messaging.ui.conversationlist.ConversationListActivity"
+            android:parentActivityName="com.android.messaging.ui.conversationlist.chats.ConversationListActivity"
             android:resizeableActivity="true"
             android:screenOrientation="user"
             android:theme="@style/Theme.Compose">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.android.messaging.ui.conversationlist.ConversationListActivity" />
+                android:value="com.android.messaging.ui.conversationlist.chats.ConversationListActivity" />
         </activity>
 
         <!-- Full-screen photo viewer -->
@@ -169,13 +169,13 @@
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:exported="true"
             android:label="@string/settings_activity_title"
-            android:parentActivityName="com.android.messaging.ui.conversationlist.ConversationListActivity"
+            android:parentActivityName="com.android.messaging.ui.conversationlist.chats.ConversationListActivity"
             android:resizeableActivity="true"
             android:screenOrientation="user"
             android:theme="@style/Theme.Compose">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.android.messaging.ui.conversationlist.ConversationListActivity" />
+                android:value="com.android.messaging.ui.conversationlist.chats.ConversationListActivity" />
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.DEFAULT" />
