+++ app/src/test/kotlin/com/android/messaging/ui/conversationlist/chats/delegate/ConversationListActionsDelegateImplTest.kt
@@ -1,12 +1,12 @@
-package com.android.messaging.ui.conversationlist.delegate
+package com.android.messaging.ui.conversationlist.chats.delegate
 
 import app.cash.turbine.test
 import com.android.messaging.data.blockedparticipants.repository.BlockedParticipantsRepository
 import com.android.messaging.data.conversation.repository.ConversationsRepository
 import com.android.messaging.data.conversationlist.repository.ConversationListRepository
 import com.android.messaging.data.conversationsettings.model.SnoozeOption
+import com.android.messaging.ui.conversationlist.chats.model.ConversationListEffect
 import com.android.messaging.ui.conversationlist.conversationItem
-import com.android.messaging.ui.conversationlist.model.ConversationListEffect
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
