+++ src/com/android/messaging/di/conversationlist/ConversationListBindsModule.kt
@@ -4,10 +4,10 @@ import com.android.messaging.data.conversationlist.repository.ConversationListRe
 import com.android.messaging.data.conversationlist.repository.ConversationListRepositoryImpl
 import com.android.messaging.data.conversationlist.store.ConversationListStatusStore
 import com.android.messaging.data.conversationlist.store.ConversationListStatusStoreImpl
+import com.android.messaging.ui.conversationlist.chats.mapper.ConversationListUiStateMapper
+import com.android.messaging.ui.conversationlist.chats.mapper.ConversationListUiStateMapperImpl
 import com.android.messaging.ui.conversationlist.mapper.ConversationListItemUiMapper
 import com.android.messaging.ui.conversationlist.mapper.ConversationListItemUiMapperImpl
-import com.android.messaging.ui.conversationlist.mapper.ConversationListUiStateMapper
-import com.android.messaging.ui.conversationlist.mapper.ConversationListUiStateMapperImpl
 import dagger.Binds
 import dagger.Module
 import dagger.Reusable
