+++ src/com/android/messaging/ui/UIIntentsImpl.java
@@ -51,7 +51,7 @@ import com.android.messaging.ui.classzero.ClassZeroActivity;
 import com.android.messaging.ui.contact.AddContactActivity;
 import com.android.messaging.ui.conversation.ConversationActivity;
 import com.android.messaging.ui.conversation.LaunchConversationActivity;
-import com.android.messaging.ui.conversationlist.ConversationListActivity;
+import com.android.messaging.ui.conversationlist.chats.ConversationListActivity;
 import com.android.messaging.ui.conversationlist.archived.ArchivedConversationListActivity;
 import com.android.messaging.ui.conversationpicker.host.forward.ForwardMessageActivity;
 import com.android.messaging.ui.conversationpicker.host.widget.WidgetPickConversationActivity;
