+++ src/com/android/messaging/ui/conversationlist/chats/ConversationListDialogs.kt
@@ -1,16 +1,16 @@
-package com.android.messaging.ui.conversationlist.ui
+package com.android.messaging.ui.conversationlist.chats
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.android.messaging.R
 import com.android.messaging.ui.common.components.SnoozeChatDialog
-import com.android.messaging.ui.conversationlist.model.ConversationListAction as Action
+import com.android.messaging.ui.conversationlist.chats.model.ConversationListAction as Action
+import com.android.messaging.ui.conversationlist.common.dialog.ConversationListDeleteDialog
 import com.android.messaging.ui.core.MessagingPreviewTheme
 
 @Composable
@@ -102,35 +102,6 @@ private fun DismissSelectionDialogsWithoutSelection(
 }
 
 @Composable
-internal fun ConversationListDeleteDialog(
-    selectedCount: Int,
-    onConfirm: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                text = pluralStringResource(
-                    id = R.plurals.delete_conversations_confirmation_dialog_title,
-                    count = selectedCount,
-                ),
-            )
-        },
-        confirmButton = {
-            TextButton(onClick = onConfirm) {
-                Text(text = stringResource(R.string.delete_conversation_confirmation_button))
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(text = stringResource(R.string.delete_conversation_decline_button))
-            }
-        },
-    )
-}
-
-@Composable
 internal fun ConversationListBlockDialog(
     destination: String,
     onConfirm: () -> Unit,
@@ -155,18 +126,6 @@ internal fun ConversationListBlockDialog(
             }
         },
     )
-}
-
-@PreviewLightDark
-@Composable
-private fun ConversationListDeleteDialogPreview() {
-    MessagingPreviewTheme {
-        ConversationListDeleteDialog(
-            selectedCount = 2,
-            onConfirm = {},
-            onDismiss = {},
-        )
-    }
 }
 
 @PreviewLightDark
