+++ src/com/android/messaging/ui/conversationlist/chats/ConversationListTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.android.messaging.ui.conversationlist.ui.topbar
+package com.android.messaging.ui.conversationlist.chats
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
@@ -14,7 +14,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.android.messaging.R
-import com.android.messaging.ui.conversationlist.model.ConversationListAction as Action
+import com.android.messaging.ui.conversationlist.chats.model.ConversationListAction as Action
 import com.android.messaging.ui.core.MessagingPreviewTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
