+++ src/com/android/messaging/ui/conversationlist/common/dialog/ConversationListDeleteDialog.kt
@@ -1,0 +1,52 @@
+package com.android.messaging.ui.conversationlist.common.dialog
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.pluralStringResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+import com.android.messaging.R
+import com.android.messaging.ui.core.MessagingPreviewTheme
+
+@Composable
+internal fun ConversationListDeleteDialog(
+    selectedCount: Int,
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit,
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = {
+            Text(
+                text = pluralStringResource(
+                    id = R.plurals.delete_conversations_confirmation_dialog_title,
+                    count = selectedCount,
+                ),
+            )
+        },
+        confirmButton = {
+            TextButton(onClick = onConfirm) {
+                Text(text = stringResource(R.string.delete_conversation_confirmation_button))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(text = stringResource(R.string.delete_conversation_decline_button))
+            }
+        },
+    )
+}
+
+@PreviewLightDark
+@Composable
+private fun ConversationListDeleteDialogPreview() {
+    MessagingPreviewTheme {
+        ConversationListDeleteDialog(
+            selectedCount = 2,
+            onConfirm = {},
+            onDismiss = {},
+        )
+    }
+}
