+++ src/com/android/messaging/ui/conversationlist/model/ConversationListContentUiState.kt
@@ -6,15 +6,6 @@ import kotlinx.collections.immutable.ImmutableSet
 import kotlinx.collections.immutable.persistentSetOf
 
 @Immutable
-internal data class ConversationListUiState(
-    val content: ConversationListContentUiState = ConversationListContentUiState.Loading,
-    val selection: ConversationListSelectionUiState = ConversationListSelectionUiState(),
-    val isScrollToTopVisible: Boolean = false,
-    val hasBlockedParticipants: Boolean = false,
-    val isDebugEnabled: Boolean = false,
-)
-
-@Immutable
 internal sealed interface ConversationListContentUiState {
 
     @Immutable
